Change project name, merge projects...
We've heard that PBHP is not intuitive for our project, and honestly when I first started this I foolishly though I was naming the "Project" that shows up at the bottom as "Punk Baby House Party" and the project names as just something that identifies which collection it is. In retrospect it is obvious I wasn't doing that, but my first pull request was the first time I had done anything on Github.

This changes the project to Punk Baby House Party, and merges the PBHP Baby Chuck into that project. I also filled in the tags right now for the next collections that will release after Cardano Gen is closed.

I also changed the name of the file so we show up alphabetically in Github.