Hopefully fix my novice mistakes
Let's hope this works...

After learning of my novice mistake regarding the copious number of policy IDs for Punk Baby House Party we have embarked upon a new era... hopefully.

This commit will move all of the current policies to live under the PBHP Cardano Gen project for now.

Meanwhile the Punk Baby House Party project has been changed to only include 3 new policy IDs and will at the end of its lifespan have a total of 6 policy IDs. Each ID corresponds to a specific series. 

The 3 series included in this commit are for Cardano Gen, Baby Chuck/Charles, and Shelley Gen which will be released at the end of September (hopefully).

Over the coming weeks we will be reaching out to all of the current owners of PBHP babies and enticing them with some sweet prizes if they swap out their babies and receive a fresh one under the new single policy ID associated with their series.

From time to time as we burn the old babies and replace them with fresh ones we will update this file to remove the no longer valid policies and hopefully delete the PBHP Cardano Gen project entirely.

If I'm still missing something essential, something that would make these changes not effective in achieving my goals, please let me know. My goal is to have the project able to function in the project filter so that people can find the NFTs and so that we will be able to use some of the various CNFT tracking tools that exist.

Thank you!