Create CARDANO CATZ
Hi there, the series 2 will be on sale as from tomorrow. Could you please verify the policies of the series 1 and series 2?
I understand that the policy of the series 1 might be an issue because it never locks, but please note that the keys have been destroyed and that no more than 126 tokens have been minted on it. The policy was owned by easycnft.art. You can verify with @vmlinuz1974, the owner of the website on twitter. In addition, it is not mentioned anywhere that the policies need to lock.

This said, could you please first of all verify the policy of the series 2? We can discuss the issue for the series 1 on a later stage.

Thanks. :)