Beast Within
When you gaze into the eyes of the beast, it will send shivers down your spine. You have a powerful beast within. You have that intensity in our eyes, that precision focus in your mind and the propensity for creation. Every human being is an immensely talented creator in different form. Remind yourself of who you are or who you can be!