Added The Cardanoverse Project Policy IDs
This is the first commit for the Cardanoverse project, where we will be minting 12 unique hand-drawn art pieces featuring the celestial bodies in our home solar system (The Sun, Mercury, Venus, Earth, our Moon, Mars, the Asteroid Belt, Jupiter, Saturn, Uranus, Neptune, and Pluto) and the background of cardano-shaped galaxies.

Note: The project shares names with another project but the IP is entirely different. In other words, the intention behind the art and the art itself are different.