Added PandemicGods
PandemicGods NFT series is based on the oilpainting series by cornelia es said of the same name. It establishes a connection between the dramas of the Greek world of Gods and our contemporary perspective. In doing so, it plays multilayered with gender identities and the competition for beauty already thematized in Greek myth.