Create SevenDeadlySins
Hello there ! (And you should respond : "General Kenobi"). Okay my bad not funny. 

I'm a crypto enthusiast and with my friend Manon,  which is the artist and the creator of the NFTs, we have created our first NFT collection with 7 unique pieces. 
The collection is about the seven deadly sins with a French Touch 🇫🇷. We would love to get verified and participate in the Cardano NFT world. 

I share with you my twitter account which is not old i never used twitter before : @TheMurphysWay or I am Murphy. 

And the link to the instagram account of my friend Manon : https://www.instagram.com/ms.tournesol/

Also, we already are the owners of the ada domain : sevendeadlysins.ada  purchased on adadomains.io     Which we are going to live as soon it will be able. 
The NFT are already minted.