added Cryptonians
These inhabitants of the virtual universe were suddenly catapulted to the real universe due to a glitch in their system, and without you they will not be able to go back home safely. Help them to reach their portals… if you are quick, you will get a reward for it!