Added CardanoRum for Verification
CardanoRum NFTs are the Digital Double for The Physical Bottles of Cardano Rum Commemorating the Alonzo Hard Fork and the 6 year anniversary of the launch of Cardano by Charles Hoskinson back in September 2015. Not to be confused with the ADA token itself which was launched two years later in September 2017. We chose one of the best rums in the world that was produced in 2015 and stored in oak casks for exactly 6 years.