Update AdaverseX
Hi 

Please review two additional Policy IDs.  One for Community (a pre-sale airdrop was given out yesterday) which will be used for the next phase of the project. 
Example CNFT from the airdrop
https://pool.pm/a947f577c0148bb7c4798dabb12242f821b99c883191a938cea6ce3c.AC5

The second is for Collector CNFTs which are sent out to members and will be use future badges, example
https://pool.pm/72973bc46ca88fbe971bacb52deea31043b1c7437c77ccabc9258302.AC1

Policy IDs displayed on the project website
https://www.adaverse-x.com/policy-ids

Kind Regards
Karl.