Create MADAzine Covers
Hey! We are from C-NFT Agency, a new project to spread the word about CNFTs projects.
We released a magazine, called MADAzine, and every issue we will mint a few copies of the cover and distribute it to the featured projects. Since it will become, with luck, a rare collectible, we want to be verified on the market.

Our socials:
https://c-nftagency.com/
https://twitter.com/UKCNFTAgent1
https://twitter.com/marco__zeroNFT