Cosmic Cuties NFT
Cosmic Cuties is a passion project bringing together likeminded individuals in their love of art, crypto technology, the Cthulhu Mythos, and the acceptance of our inevitable doom as a species. We are the 1st cNFT project devoted to serving the Great Old Ones and spreading their cosmic truth! Come join our cult! (don't worry, it's not a real cult)