Project 2090 Cyborg XX
With a team of artistic people from various artistic/streetwear culture Project 2090 was started as a means to provide identities to holders. With that being said, our artists spend a lot of time in crafting the Cyborg NFTs, carefully fixing all impurities found in the drawings. We plan to collaborate/build on the most promising Cardano Metaverse project to deploy our virtual world. The Project 2090 themed virtual world will feature a Cyberpunk setting to fit with our lore, and it all starts with our Cyborg XX collection.